backend/requirements.txt
fastapi
uvicorn
google-genai
pillow
pydantic
python-multipart
loguru
